Makefile

+++ window.c
@@ -1,12 +1,23 @@
 #include "window.h"
 
+int position = 0;
+
 void clean_term(int height, int width, int startx) {
 	mvprintw(height, startx, " > ");
 	for (int i=startx + 3; i<width-3; i++) 
 		mvprintw(height, i, " ");	
 }
 
-void term_win(int height, int width, int starty, int startx, int* cur) {
+void check_ret(int ret, int height, int startx) {
+	if(ret==-1)
+		mvprintw(height, startx + 4, "Errore nell' invio del segnale al processo");
+  else 
+   	mvprintw(height, startx + 4, "Segnale inviato correttamente");
+  refresh();
+  sleep(3);
+}
+
+void term_win(int height, int width, int starty, int startx, int* cur, int* active_proc) {
 	WINDOW* input = create_newwin(1, width, height, startx, 0);
 	mvprintw(height, startx, " > ");
   int ch;
@@ -14,8 +25,6 @@ void term_win(int height, int width, int starty, int startx, int* cur) {
 	move(height, startx + 4 + *cur);
 	while(1) {
 		ch = getch();
-		//getstr(str);
-		//printw("Hai inserito: %d", (int) ch);
 		if (ch == 263) {		//backspace
 			if (*cur != 0) {
 				*cur -= 1;
@@ -23,55 +32,58 @@ void term_win(int height, int width, int starty, int startx, int* cur) {
 				buf[*cur] = ' ';
 			}
 		}
+		else if (ch == KEY_DOWN) {
+			if (position < *active_proc - 32)
+				position += 1;
+		}
+		else if (ch == KEY_UP) {
+			if (position > 0)
+				position -= 1;		
+		}
 		else if (ch == 10) { // enter
 			clean_term(height, width, startx);
 			buf[*cur] = '\0';
 			*cur = 0;
 			char* command = strtok(buf, " ");
+			pid_t pid;
 			if (strcmp(command, SHUTDOWN) == 0) {
-				mvprintw(height, startx + 4, "Adioss");
-				refresh();
 				break;
 			}
 			else if (strcmp(command, KILL) == 0) {
-				pid_t pid = atoi(strtok(NULL, " "));
-				int ret=kill(pid, SIGTERM);
-      	if(ret==-1)
-					mvprintw(height, startx + 4, "Errore nell' invio del segnale al processo");
-        else 
-        	mvprintw(height, startx + 4, "Segnale inviato correttamente");
-        refresh();
-        sleep(3);
+				command = strtok(NULL, " ");
+				pid = atoi(command);
+				// TODO verifica che inviando 'kill' non fa segmentation fault
+//				mvprintw(height, startx + 4 , "%s", command);
+	//			refresh();
+		//		sleep(2);
+				if (pid != 0) {
+					int ret=kill(pid, SIGTERM);
+  	    	check_ret(ret, height, startx);
+				}
 			}
 			else if (strcmp(command, RESUME) == 0) {
-				pid_t pid = atoi(strtok(NULL, " "));
-				int ret=kill(pid, SIGCONT);
-      	if(ret==-1)
-					mvprintw(height, startx + 4, "Errore nell' invio del segnale al processo");
-        else 
-        	mvprintw(height, startx + 4, "Segnale inviato correttamente");
-        refresh();
-        sleep(3);
+				command = strtok(NULL, " ");
+				pid = atoi(command);
+				if (pid != 0) {
+					int ret=kill(pid, SIGCONT);
+  	    	check_ret(ret, height, startx);
+				}
 			}
 			else if (strcmp(command, SUSPEND) == 0) {
-				pid_t pid = atoi(strtok(NULL, " "));
-				int ret=kill(pid, SIGSTOP);
-      	if(ret==-1)
-					mvprintw(height, startx + 4, "Errore nell' invio del segnale al processo");
-        else 
-        	mvprintw(height, startx + 4, "Segnale inviato correttamente");
-        refresh();
-        sleep(3);
+				command = strtok(NULL, " ");
+				pid = atoi(command);
+				if (pid != 0) {
+					int ret=kill(pid, SIGSTOP);
+  	    	check_ret(ret, height, startx);
+				}
 			}
 			else if (strcmp(command, QUIT) == 0) {
-				pid_t pid = atoi(strtok(NULL, " "));
-				int ret=kill(pid, SIGQUIT);
-      	if(ret==-1)
-					mvprintw(height, startx + 4, "Errore nell' invio del segnale al processo");
-        else 
-        	mvprintw(height, startx + 4, "Segnale inviato correttamente");
-        refresh();
-        sleep(3);
+				command = strtok(NULL, " ");
+				pid = atoi(command);
+				if (pid != 0) {
+					int ret=kill(pid, SIGQUIT);
+  	    	check_ret(ret, height, startx);
+				}
 			}
 			clean_term(height, width, startx);
 		}
@@ -132,33 +144,33 @@ void print_main_win(WINDOW* win, Process* proc, Global* glob, int height, int wi
 	sprintf(tmp, "%d", glob->zombieProc);
 	mvprintw(starty + 2, 11*(width-startx)/12 - (strlen(tmp)+7), "%s zombie", tmp);
 
-	mvprintw(starty, (width-startx)/2 - strlen(" My TOP ")/2, " My TOP ");
-	for (int i=startx + 1; i<(width-startx); i++) {
+	for (int i=startx + 5; i<(width-startx); i++) {
 		for (int j=starty+10; j<height-3; j++)
 			mvprintw(starty + j, i, " ");
 	}
 	for (int p=0; p<height - 14; p++) {
-		mvprintw(starty + 11 + p, 1*(width-startx)/24 - strlen(proc[p].pid)/2, "%s", proc[p].pid);
-		mvprintw(starty + 11 + p, 3*(width-startx)/24 - strlen(proc[p].user)/2, "%s", proc[p].user);
-		sprintf(tmp, "%d", proc[p].priority);
+		Process process = proc[p + position];
+		mvprintw(starty + 11 + p, 1*(width-startx)/24 - strlen(process.pid)/2, "%s", process.pid);
+		mvprintw(starty + 11 + p, 3*(width-startx)/24 - strlen(process.user)/2, "%s", process.user);
+		sprintf(tmp, "%d", process.priority);
 		mvprintw(starty + 11 + p, 5*(width-startx)/24 - strlen(tmp)/2, "%s", tmp);
-		sprintf(tmp, "%d", proc[p].ni);
+		sprintf(tmp, "%d", process.ni);
 		mvprintw(starty + 11 + p, 7*(width-startx)/24 - strlen(tmp)/2, "%s", tmp);
-		sprintf(tmp, "%ld", proc[p].virt);
+		sprintf(tmp, "%ld", process.virt);
 		mvprintw(starty + 11 + p, 9*(width-startx)/24 - strlen(tmp)/2, "%s", tmp);
-		sprintf(tmp, "%ld", proc[p].res);
+		sprintf(tmp, "%ld", process.res);
 		mvprintw(starty + 11 + p, 11*(width-startx)/24 - strlen(tmp)/2, "%s", tmp);
-		sprintf(tmp, "%ld", proc[p].shr);
+		sprintf(tmp, "%ld", process.shr);
 		mvprintw(starty + 11 + p, 13*(width-startx)/24 - strlen(tmp)/2, "%s", tmp);
-		mvprintw(starty + 11 + p, 15*(width-startx)/24, "%c", proc[p].status);
-		sprintf(tmp, "%0.1f", proc[p].cpu);
+		mvprintw(starty + 11 + p, 15*(width-startx)/24, "%c", process.status);
+		sprintf(tmp, "%0.1f", process.cpu);
 		mvprintw(starty + 11 + p, 17*(width-startx)/24 - strlen(tmp)/2, "%s", tmp);
-		sprintf(tmp, "%0.1f", proc[p].mem);
+		sprintf(tmp, "%0.1f", process.mem);
 		mvprintw(starty + 11 + p, 19*(width-startx)/24 - strlen(tmp)/2, "%s", tmp);
-		sprintf(tmp, "%0.3f", proc[p].time);
+		sprintf(tmp, "%0.3f", process.time);
 		mvprintw(starty + 11 + p, 21*(width-startx)/24 - strlen(tmp)/2, "%s", tmp);
-		proc[p].command[(width-startx)/12 - 2] = '\0';
-		mvprintw(starty + 11 + p, 23*(width-startx)/24 - strlen(proc[p].command)/2, "%s", proc[p].command);
+		process.command[(width-startx)/12 - 2] = '\0';
+		mvprintw(starty + 11 + p, 23*(width-startx)/24 - strlen(process.command)/2, "%s", process.command);
 	}
 	move(height, startx + 4 + (*cur));
 	refresh();
@@ -174,6 +186,14 @@ WINDOW* create_main_win(int* height, int* width, int* starty, int* startx) {
   *starty = 0;	/* Calculating for a center placement */
   *startx = 0;	/* of the window		*/
   my_win = create_newwin(*height, *width, *starty, *startx, 1);
+ 
+  if (has_colors()){
+    start_color();
+    
+    init_pair(1, COLOR_WHITE, COLOR_BLACK);
+    bkgd(COLOR_PAIR(1));
+  }
+
   return my_win;
 }
 
@@ -181,7 +201,7 @@ WINDOW* create_main_win(int* height, int* width, int* starty, int* startx) {
 WINDOW *create_newwin(int height, int width, int starty, int startx, int box) {	
     WINDOW *local_win;
     local_win = newwin(height, width, starty, startx);
-    if (box) // solo se box è 1
+    if (box)
 	    box(local_win, 0, 0);		/* 0, 0 gives default characters 
 					   for the vertical and horizontal
 					    lines			*/
@@ -189,23 +209,8 @@ WINDOW *create_newwin(int height, int width, int starty, int startx, int box) {
     return local_win;
 }
 
-void destroy_win(WINDOW *local_win) {
-    	/* box(local_win, ' ', ' '); : This won't produce the desired
-  	 * result of erasing the window. It will leave it's four corners 
-	 * and so an ugly remnant of window.  */
-    
+void destroy_win(WINDOW *local_win) {    
     wborder(local_win, ' ', ' ', ' ',' ',' ',' ',' ',' ');
-
-	/* The parameters taken are 
-	 * 1. win: the window on which to operate
-	 * 2. ls: character to be used for the left side of the window 
-	 * 3. rs: character to be used for the right side of the window 
-	 * 4. ts: character to be used for the top side of the window 
-	 * 5. bs: character to be used for the bottom side of the window 
- 	 * 6. tl: character to be used for the top left corner of the window 
-	 * 7. tr: character to be used for the top right corner of the window 
-	 * 8. bl: character to be used for the bottom left corner of the window 
-	 * 9. br: character to be used for the bottom right corner of the window */
 
     wrefresh(local_win);
     delwin(local_win);
